requirements.txt
antiorm==1.2.1
blinker==1.8.2
click==8.1.7
db==0.1.1
db-sqlite3==0.0.1
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==1.0.1
Werkzeug==3.0.3
